requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
exceptiongroup==1.0.0
factory-boy==3.2.1
Faker==12.0.1
freezegun==1.2.2
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
kombu==5.2.4
mixer==7.2.2
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
